requirements.dev.txt
pandas
numpy
jupyter
rake_nltk
wordcloud
frozendict
ipywidgets==7.7.1
spacy
scikit-learn
colorlog
plotly
pyldavis==3.4.0
transformers
pyarrow
coolname
